Move all serialization and deserialization into encoding module.
